listing_to_video_service/requirements.txt
fastapi
uvicorn
openai
python-dotenv
firebase-admin
elevenlabs
remotion
python-multipart
